refactor: rename washing commands namespace
