+++ backend/src/controlmat.Application/Common/Commands/WashCycle/AddProtCommand.cs
@@ -5,7 +5,7 @@ using Controlmat.Domain.Entities;
 using Controlmat.Domain.Interfaces;
 using System;
 
-namespace Controlmat.Application.Common.Commands.Washing;
+namespace Controlmat.Application.Common.Commands.WashCycle;
 
 public static class AddProtCommand
 {
